requirements.txt
future
redis
emoji
pydub
numpy
multicolorcaptcha
redis==4.1.1
tzlocal==2.0
python-barcode==0.13.1
pykeyboard
lxml
wget
PyDictionary
mtranslate
googletrans
aioredis==2.0.0
gpytranslate
gtts
python-telegram-bot==13.11
faker
markdown
beautifulsoup4
opencv-python-headless
requests==2.27.1
sqlalchemy==1.3.23
psycopg2-binary
feedparser
hachoir
pynewtonmath
spongemock
zalgo-text
geopy
youtube_search
youtube-dl
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
google_trans_new
jikanpy
coffeehouse==2.2.4
regex
bleach
git+https://github.com/python-telegram-bot/ptbcontrib.git
wikipedia
telethon==1.24.0
telegraph
heroku3
spamwatch
alphabet_detector
pybase64
pySmartDL
validators
nekos.py==1.0.3
aiofiles
pyrate-limiter
cachetools
ujson
pretty_errors
TgCrypto
Pyrogram==1.4.9
PyYAML>=5.3.1
pretty_errors

futures==3.1.1
ffmpeg==1.4
ffmpeg-python
markdown2
youtube-search-python
asyncio
dateparser==1.0.0
pymongo==3.12.1
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
fontTools
bing_image_downloader
search_engine_parser
pytz
lyricsgenius
tswift
envparse
better_profanity
nudepy
motor
cloudscraper
python-arq
pytube
tldextract
cloudmersive_virus_api_client
youtube_search
pretty-errors
odmantic
pyrogram
beautifulsoup4==4.10.0
bidict==0.21.3
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
comtypes==1.1.10
decorator==5.1.0
future==0.18.2
geocoder==1.38.1
idna==3.2
jsonlib-python3==1.6.1
loc==0.0.2

PyRect==0.1.4
PyScreeze==0.1.28
pyttsx3==2.90
pytube==11.0.1
pytweening==1.0.4

ratelim==0.1.6
requests
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
urllib3==1.26.7
wikipedia==1.4.0
